backend/requirements.txt
APScheduler==3.9.1
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.3
django-appconf==1.0.5
django-cors-headers==3.11.0
django-extensions==3.1.5
django-imagekit==4.1.0
django-sslserver==0.22
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy==1.22.3
packaging==21.3
pandas==1.4.1
pilkit==2.0
Pillow==9.0.1
PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scipy==1.8.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.9
